+++ gymnax/environments/__init__.py
@@ -32,6 +32,8 @@ from .misc import (
     MetaMaze,
     PointRobot,
     Reacher,
+    Swimmer,
+    Pong,
 )
 
 
@@ -61,4 +63,6 @@ __all__ = [
     "MetaMaze",
     "PointRobot",
     "Reacher",
+    "Swimmer",
+    "Pong",
 ]